+++ examples/data-sources/graviteeam_user_credentials/data-source.tf
@@ -1,0 +1,4 @@
+data "graviteeam_user_credentials" "example" {
+  domain_id = graviteeam_domain.example.id
+  user_id   = graviteeam_user.example.id
+}